update :: 중복되는 코드를 util로 처리
